requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
flask-marshmallow
flask_cors
marshmallow-sqlalchemy
flask-jwt-extended
pymysql
cryptography
Werkzeug
openai
python-dotenv